Add BaseScriptTestCase and update the unit tests to use it.
In addition to simplifying the unit tests, this has code that supports
both ts_salobj 5.4/ts_xml 4.7 and earlier ts_salobj with ts_xml 4.6.

In addition I modified the flake8 configuration to change W504 to W503
(binary operators should be at the start of a continuation line)
and reformatted the unit tests with ``black``.