Implement latest PR review comments.
On the enable and standby group add a snippet that will allow the schema to output what the values of the `ignore` field can be. They are now all getting the information from the group itself, making the schema_yaml a f-string and adding `self.group.components` to the string.

Removed empty docstring sections.