In ``base_track_target_and_take_image.py``, add ``az_wrap_strategy`` to the script configuration.
This allows users to specify the azimuth wrap strategy the TCS should use when slewing to a target.
The parameter is exposed as an enumeration with all the available options.
Users select an option by adding one of the available strings.
When configuring the Script, the ``configure`` method will convert the string into the appropriate enumeration, calling in the ``tcs`` property to return the ``WrapStrategy`` enumeration.