In ``maintel/track_target_and_take_image_gencam.py``, update ``get_schema`` method to stop deleting ``band_filter`` from the required configuration attributes.
Previously we thought it would be ok to remove this attribute from the configuration since the generic cameras, which this script is designed to work with, don't necessarily have a filter wheel or instrument configuration.
But this oversight doesn't take into account the fact that this Script is designed to work with the Scheduler and, for this type of Script, we can not remove any of the basic set of required parameters.
Adding new parameters is ok though.

If calling this script from the script queue one can simply pass in an empty string for ``band_filter``.
But, keep in mind this one in particular is designed to work with the Scheduler.