Add new ``SystemWideShutdown`` script.
This script is designed to help shutdown the system for upgrades or general maintenance.
When executed it will search for all running components and then proceed to send them to offline.
The discovery process consists of getting the list of CSCs from the idl directory and then trying to find running CSCs by their heartbeats.
It also discovers indexed CSCs by using remotes with index=0 and accumulating heartbeats over a period of time.